requirements.txt
tqdm
h5py==2.10.0
hdf5storage==0.1.15
jsonpatch==1.24
jsonpointer==2.0
matplotlib==3.1.1
numpy==1.17.3
Pillow==6.2.1
pyparsing==2.4.5
scipy==1.3.2
torch==1.3.1
torchfile==0.1.0
torchvision==0.4.2
visdom==0.1.8.9
websocket-client==0.56.0
